# Data-Gap-AI/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ColumnInfo {
  name: string;
  type: string;
  sampleValues: string[];
  nullPercent: number;
  relevanceScore: number;
  relevanceReason: string;
}

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type AnalysisRiskLevel = typeof AnalysisRiskLevel[keyof typeof AnalysisRiskLevel] | null;


export const AnalysisRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Analysis {
  id: number;
  question: string;
  filename: string;
  rowCount: number;
  columnCount: number;
  status: AnalysisStatus;
  /** @nullable */
  readinessScore: number | null;
  /** @nullable */
  riskLevel: AnalysisRiskLevel;
  createdAt: string;
}

export interface AnalysisInput {
  /** @minLength 5 */
  question: string;
  filename: string;
  /** Raw CSV content as a string */
  csvData: string;
}

export type DataItemImportance = typeof DataItemImportance[keyof typeof DataItemImportance];


export const DataItemImportance = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface DataItem {
  name: string;
  description: string;
  importance: DataItemImportance;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type RecommendationEffort = typeof RecommendationEffort[keyof typeof RecommendationEffort];


export const RecommendationEffort = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Recommendation {
  priority: RecommendationPriority;
  action: string;
  impact: string;
  effort: RecommendationEffort;
}

export type TimelineEventStatus = typeof TimelineEventStatus[keyof typeof TimelineEventStatus];


export const TimelineEventStatus = {
  required: 'required',
  optional: 'optional',
  future: 'future',
} as const;

export interface TimelineEvent {
  phase: string;
  description: string;
  duration: string;
  status: TimelineEventStatus;
}

export type AnalysisReportRiskLevel = typeof AnalysisReportRiskLevel[keyof typeof AnalysisReportRiskLevel];


export const AnalysisReportRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface AnalysisReport {
  id: number;
  analysisId: number;
  readinessScore: number;
  riskLevel: AnalysisReportRiskLevel;
  availableData: DataItem[];
  missingData: DataItem[];
  recommendations: Recommendation[];
  explanation: string;
  unansweredQuestions: string[];
  columnAnalysis: ColumnInfo[];
  createdAt: string;
}

export interface DemoData {
  analysis: Analysis;
  report: AnalysisReport;
}

export interface AnalysisStats {
  totalAnalyses: number;
  avgReadinessScore: number;
  highRiskCount: number;
  mediumRiskCount: number;
  lowRiskCount: number;
  recentAnalyses: Analysis[];
}