# Data-Gap-AI/lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRiskLevel } from './analysisRiskLevel';
import type { AnalysisStatus } from './analysisStatus';

export interface Analysis {
  id: number;
  question: string;
  filename: string;
  rowCount: number;
  columnCount: number;
  status: AnalysisStatus;
  /** @nullable */
  readinessScore: number | null;
  /** @nullable */
  riskLevel: AnalysisRiskLevel;
  createdAt: Date;
}