# Data-Gap-AI/lib/api-zod/src/generated/types/analysisInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisInput {
  /** @minLength 5 */
  question: string;
  filename: string;
  /** Raw CSV content as a string */
  csvData: string;
}