# Data-Gap-AI/lib/api-zod/src/generated/types/analysisReport.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisReportRiskLevel } from './analysisReportRiskLevel';
import type { ColumnInfo } from './columnInfo';
import type { DataItem } from './dataItem';
import type { Recommendation } from './recommendation';

export interface AnalysisReport {
  id: number;
  analysisId: number;
  readinessScore: number;
  riskLevel: AnalysisReportRiskLevel;
  availableData: DataItem[];
  missingData: DataItem[];
  recommendations: Recommendation[];
  explanation: string;
  unansweredQuestions: string[];
  columnAnalysis: ColumnInfo[];
  createdAt: Date;
}