# Data-Gap-AI/lib/api-zod/src/generated/types/analysisStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Analysis } from './analysis';

export interface AnalysisStats {
  totalAnalyses: number;
  avgReadinessScore: number;
  highRiskCount: number;
  mediumRiskCount: number;
  lowRiskCount: number;
  recentAnalyses: Analysis[];
}