# Data-Gap-AI/lib/api-zod/src/generated/types/columnInfo.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ColumnInfo {
  name: string;
  type: string;
  sampleValues: string[];
  nullPercent: number;
  relevanceScore: number;
  relevanceReason: string;
}