# Data-Gap-AI/lib/api-zod/src/generated/types/dataItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DataItemImportance } from './dataItemImportance';

export interface DataItem {
  name: string;
  description: string;
  importance: DataItemImportance;
}