# Data-Gap-AI/lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DataGap AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineEventStatus } from './timelineEventStatus';

export interface TimelineEvent {
  phase: string;
  description: string;
  duration: string;
  status: TimelineEventStatus;
}